Move get_authority_code to utils and add tests for this function
