+++ tests/test_water_depth_difference_algorithm.py
@@ -1,9 +1,10 @@
 import pytest
 import tempfile
 from pathlib import Path
-from osgeo import gdal, osr
+from osgeo import gdal
 import numpy as np
 from threedi_results_analysis import PLUGIN_DIR
+from threedi_results_analysis.tests.utilities import create_test_raster as create_test_raster_full
 import threedi_results_analysis.processing.water_depth_difference_algorithm as wdda
 
 RASTER_DIR = PLUGIN_DIR / "tests" / "data" / "rasters"
@@ -15,48 +16,18 @@ OFFSET_X, OFFSET_Y = 140000, 460000
 
 def create_test_raster(filename, array_data, origin_x=OFFSET_X, origin_y=OFFSET_Y, pixel_width=10, pixel_height=-10,
                        epsg_code=28992, no_data_value=-9999, x_skew=0, y_skew=0):
-    """
-    Create a test GeoTIFF raster file with given parameters
-
-    Parameters:
-    filename - output file path
-    array_data - 2D numpy array with raster data
-    origin_x, origin_y - coordinates of the upper left corner
-    pixel_width, pixel_height - pixel size (height is usually negative)
-    epsg_code - projection code (default: 28992 - Dutch RD New)
-    """
     output_path = str(Path(TMP_DIR.name) / filename)
-    # Get dimensions
-    rows, cols = array_data.shape
-
-    # Create driver
-    driver = gdal.GetDriverByName('GTiff')
-
-    # Create dataset with 1 band
-    out_raster = driver.Create(
-        output_path,
-        cols, rows, 1,
-        gdal.GDT_Float32,
-        options=["COMPRESS=DEFLATE", "PREDICTOR=2", "TILED=YES"]
-    )
-
-    # Set geotransform (upper_left_x, pixel_width, x_skew, upper_left_y, y_skew, pixel_height)
-    out_raster.SetGeoTransform((origin_x, pixel_width, x_skew, origin_y, y_skew, pixel_height))
-
-    # Set projection
-    srs = osr.SpatialReference()
-    srs.ImportFromEPSG(epsg_code)
-    out_raster.SetProjection(srs.ExportToWkt())
-
-    # Write data
-    band = out_raster.GetRasterBand(1)
-    band.SetNoDataValue(no_data_value)  # Set NoData value
-    band.WriteArray(array_data)
-
-    # Clean up
-    band.FlushCache()
-    out_raster = None
-    return str(output_path)
+    return create_test_raster_full(
+        output_path=output_path,
+        array_data=array_data,
+        origin_x=origin_x,
+        origin_y=origin_y,
+        pixel_width=pixel_width,
+        pixel_height=pixel_height,
+        epsg_code=epsg_code,
+        no_data_value=no_data_value,
+        x_skew=x_skew,
+        y_skew=y_skew)
 
 
 def get_raster(array_data, origin_x, origin_y, no_data_value=-9999):
